+++ CHANGELOG.md
@@ -5,7 +5,50 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 1.4.0
+
+### Added
+
+- Python scripts / jobs
+  - Curious invitations to REDCap
+
+## 1.3.0
+
+### Added
+
+- Python scripts / jobs
+  - Curious alerts to REDCap
+
+### Updated
+
+- initial Terraform configuration
+- utility function library
+
+## 1.2.0
+
+### Added
+
+- Python scripts / jobs
+  - REDCap to Curious
+
+### Updated
+
+- initial Terraform configuration
+- utility function library
+
+## 1.1.0
+
+### Added
+
+- Python scripts / jobs
+  - REDCap to REDCap
+
+### Updated
+
+- initial Terraform configuration
+- utility function library
+
+## 1.0.0
 
 ### Added
 
@@ -13,6 +56,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - utility function library
 - Python scripts / jobs
   - Ripple to REDCap
-  - REDCap to REDCap
-  - REDCap to Curious
-  - Curious alerts to REDCap